run pg test with normal image and sql tests with full image

+++ docker/integration-tests/run-integration-tests.sh
@@ -27,7 +27,7 @@ echo "tagToTest          = $tagToTest"
 echo "testVersion        = $testVersion"
 echo "version            = $version"
 
-
+# build to orthanc-under-tests image
 docker build --build-arg IMAGE_TAG=$tagToTest -f orthanc-under-tests/Dockerfile -t orthanc-under-tests orthanc-under-tests
 
 pushd ../..  # we need to be at 'root' to use bash-helpers !
@@ -44,7 +44,7 @@ popd  # back to docker/integration-tests folder
 
 
 ############ run NewTests first
-
+rm -rf orthanc-tests-repo/
 hg clone https://hg.orthanc-server.com/orthanc-tests/ -r $orthanc_tests_revision orthanc-tests-repo
 
 pushd orthanc-tests-repo/NewTests
@@ -66,55 +66,61 @@ python3 -u main.py --pattern=* \
 popd
 ############ run NewTests
 
+############ run legacy tests
 
+if [[ $image == "normal" ]]; then
 
-docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests -t orthanc-tests orthanc-tests
-docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-dicomweb -t orthanc-tests-dicomweb orthanc-tests
-docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-worklists -t orthanc-tests-worklists orthanc-tests
-docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-recycling -t orthanc-tests-recycling orthanc-tests
-docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-transfers -t orthanc-tests-transfers orthanc-tests
-docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-wsi -t orthanc-tests-wsi orthanc-tests
-docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-webdav -t orthanc-tests-webdav orthanc-tests
-docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-cget -t orthanc-tests-cget orthanc-tests
+    docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests -t orthanc-tests orthanc-tests
+    docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-dicomweb -t orthanc-tests-dicomweb orthanc-tests
+    docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-worklists -t orthanc-tests-worklists orthanc-tests
+    docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-recycling -t orthanc-tests-recycling orthanc-tests
+    docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-transfers -t orthanc-tests-transfers orthanc-tests
+    docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-wsi -t orthanc-tests-wsi orthanc-tests
+    docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-webdav -t orthanc-tests-webdav orthanc-tests
+    docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests-cget -t orthanc-tests-cget orthanc-tests
 
-COMPOSE_FILE=docker-compose.sqlite.yml                      docker-compose down -v
-COMPOSE_FILE=docker-compose.sqlite.yml                      docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
+    COMPOSE_FILE=docker-compose.sqlite.yml                      docker-compose down -v
+    COMPOSE_FILE=docker-compose.sqlite.yml                      docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
 
-COMPOSE_FILE=docker-compose.dicomweb.yml                    docker-compose down -v
-COMPOSE_FILE=docker-compose.dicomweb.yml                    docker-compose up --build --exit-code-from orthanc-tests-dicomweb --abort-on-container-exit
+    COMPOSE_FILE=docker-compose.dicomweb.yml                    docker-compose down -v
+    COMPOSE_FILE=docker-compose.dicomweb.yml                    docker-compose up --build --exit-code-from orthanc-tests-dicomweb --abort-on-container-exit
 
-COMPOSE_FILE=docker-compose.postgres.yml                    docker-compose down -v
-COMPOSE_FILE=docker-compose.postgres.yml                    docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
+    COMPOSE_FILE=docker-compose.postgres.yml                    docker-compose down -v
+    COMPOSE_FILE=docker-compose.postgres.yml                    docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
 
-COMPOSE_FILE=docker-compose.odbc-postgres.yml               docker-compose down -v
-COMPOSE_FILE=docker-compose.odbc-postgres.yml               docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
+    COMPOSE_FILE=docker-compose.odbc-postgres.yml               docker-compose down -v
+    COMPOSE_FILE=docker-compose.odbc-postgres.yml               docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
 
-COMPOSE_FILE=docker-compose.odbc-sqlite.yml                 docker-compose down -v
-COMPOSE_FILE=docker-compose.odbc-sqlite.yml                 docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
+    COMPOSE_FILE=docker-compose.odbc-sqlite.yml                 docker-compose down -v
+    COMPOSE_FILE=docker-compose.odbc-sqlite.yml                 docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
 
-COMPOSE_FILE=docker-compose.mysql.yml                       docker-compose down -v
-COMPOSE_FILE=docker-compose.mysql.yml                       docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
+    COMPOSE_FILE=docker-compose.mysql.yml                       docker-compose down -v
+    COMPOSE_FILE=docker-compose.mysql.yml                       docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
 
-COMPOSE_FILE=docker-compose.webdav.yml                      docker-compose down -v
-COMPOSE_FILE=docker-compose.webdav.yml                      docker-compose up --build --exit-code-from orthanc-tests-webdav --abort-on-container-exit
+    COMPOSE_FILE=docker-compose.webdav.yml                      docker-compose down -v
+    COMPOSE_FILE=docker-compose.webdav.yml                      docker-compose up --build --exit-code-from orthanc-tests-webdav --abort-on-container-exit
 
-COMPOSE_FILE=docker-compose.cget.yml                        docker-compose down -v
-COMPOSE_FILE=docker-compose.cget.yml                        docker-compose up --build --exit-code-from orthanc-tests-cget --abort-on-container-exit
+    COMPOSE_FILE=docker-compose.cget.yml                        docker-compose down -v
+    COMPOSE_FILE=docker-compose.cget.yml                        docker-compose up --build --exit-code-from orthanc-tests-cget --abort-on-container-exit
 
-COMPOSE_FILE=docker-compose.s3.yml                         docker-compose down -v
-COMPOSE_FILE=docker-compose.s3.yml                         docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
+    COMPOSE_FILE=docker-compose.s3.yml                         docker-compose down -v
+    COMPOSE_FILE=docker-compose.s3.yml                         docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
 
-COMPOSE_FILE=docker-compose.wsi.yml                         docker-compose down -v
-COMPOSE_FILE=docker-compose.wsi.yml                         docker-compose up --build --exit-code-from orthanc-tests-wsi --abort-on-container-exit
+    COMPOSE_FILE=docker-compose.wsi.yml                         docker-compose down -v
+    COMPOSE_FILE=docker-compose.wsi.yml                         docker-compose up --build --exit-code-from orthanc-tests-wsi --abort-on-container-exit
 
-COMPOSE_FILE=docker-compose.transfers.yml                   docker-compose down -v
-COMPOSE_FILE=docker-compose.transfers.yml                   docker-compose up --build --exit-code-from orthanc-tests-transfers --abort-on-container-exit
+    COMPOSE_FILE=docker-compose.transfers.yml                   docker-compose down -v
+    COMPOSE_FILE=docker-compose.transfers.yml                   docker-compose up --build --exit-code-from orthanc-tests-transfers --abort-on-container-exit
 
-COMPOSE_FILE=docker-compose.recycling.yml                   docker-compose down -v
-COMPOSE_FILE=docker-compose.recycling.yml                   docker-compose up --build --exit-code-from orthanc-tests-recycling --abort-on-container-exit
+    COMPOSE_FILE=docker-compose.recycling.yml                   docker-compose down -v
+    COMPOSE_FILE=docker-compose.recycling.yml                   docker-compose up --build --exit-code-from orthanc-tests-recycling --abort-on-container-exit
 
-COMPOSE_FILE=docker-compose.worklists.yml                   docker-compose down -v
-COMPOSE_FILE=docker-compose.worklists.yml                   docker-compose up --build --exit-code-from orthanc-tests-worklists --abort-on-container-exit
+    COMPOSE_FILE=docker-compose.worklists.yml                   docker-compose down -v
+    COMPOSE_FILE=docker-compose.worklists.yml                   docker-compose up --build --exit-code-from orthanc-tests-worklists --abort-on-container-exit
+
+# note: not functional yet:
+# COMPOSE_FILE=docker-compose.odbc-mysql.yml               docker-compose down -v
+# COMPOSE_FILE=docker-compose.odbc-mysql.yml               docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
 
 # TODO: add tests:
 # - CheckScuTranscoding.py
@@ -122,13 +128,13 @@ COMPOSE_FILE=docker-compose.worklists.yml                   docker-compose up --
 # - CheckHttpServerSecurity.py
 # - CheckDicomTls.py
 
+else  # full images (MSSQL only !)
+
+    docker build --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests -t orthanc-tests orthanc-tests
+
+    COMPOSE_FILE=docker-compose.odbc-sql-server.yml             docker-compose down -v
+    COMPOSE_FILE=docker-compose.odbc-sql-server.yml             docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
+
+fi
 
 
-
-# TODO: reenable sqlserver odbc drivers
-# COMPOSE_FILE=docker-compose.odbc-sql-server.yml          docker-compose down -v
-# COMPOSE_FILE=docker-compose.odbc-sql-server.yml          docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
-
-# note: not functional yet:
-# COMPOSE_FILE=docker-compose.odbc-mysql.yml               docker-compose down -v
-# COMPOSE_FILE=docker-compose.odbc-mysql.yml               docker-compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
